requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
pyYAML
evidently
matplotlib
dill
imblearn
opendatasets
-e .